Add comments and refactors to the code
