+++ src/types/types.ts
@@ -1,0 +1,3 @@
+export type Turns = '❌' | '⚪';
+
+export type PlayerOrIa = 'player' | 'ia';
