+++ src/utils/getWinner.ts
@@ -1,5 +1,5 @@
-import { Turns } from '../components/Board';
 import { WINNING_COMBINATIONS } from '../constants/constants';
+import { Turns } from '../types/types';
 
 export const getWinner = (newBoard: (null | Turns)[]) => {
     for (const combination of WINNING_COMBINATIONS) {
